attempt to setup sentry/nextjs
