+++ .gitignore
@@ -40,3 +40,5 @@ cypress.env.json
 /cypress/videos
 /cypress/screenshots
 /cypress/integration/examples
+
+.sentryclirc
