+++ next.config.js
@@ -4,30 +4,8 @@ const withBundleAnalyzer = require("@next/bundle-analyzer")({
 });
 const { i18n } = require("./next-i18next.config");
 
-// Use the SentryWebpack plugin to upload the source maps during build step
-const SentryWebpackPlugin = require("@sentry/webpack-plugin");
+const { withSentryConfig } = require("@sentry/nextjs");
 
-const {
-  NEXT_PUBLIC_SENTRY_DSN: SENTRY_DSN,
-  SENTRY_ORG,
-  SENTRY_PROJECT,
-  SENTRY_AUTH_TOKEN,
-  NODE_ENV,
-  VERCEL_GIT_COMMIT_SHA,
-  VERCEL_GITHUB_COMMIT_SHA,
-  VERCEL_GITLAB_COMMIT_SHA,
-  VERCEL_BITBUCKET_COMMIT_SHA,
-  SOURCE_VERSION,
-} = process.env;
-
-const COMMIT_SHA =
-  VERCEL_GIT_COMMIT_SHA ||
-  VERCEL_GITHUB_COMMIT_SHA ||
-  VERCEL_GITLAB_COMMIT_SHA ||
-  VERCEL_BITBUCKET_COMMIT_SHA ||
-  SOURCE_VERSION;
-
-process.env.SENTRY_DSN = SENTRY_DSN;
 const basePath = "";
 
 const scheme =
@@ -45,57 +23,12 @@ if (process.env.NEXT_PUBLIC_VERCEL_ENV === "preview") {
 const hasAssetPrefix =
   process.env.ASSET_PREFIX !== "" && process.env.ASSET_PREFIX !== undefined;
 
-module.exports = withPlugins([[withBundleAnalyzer]], {
+const moduleExports = withPlugins([[withBundleAnalyzer]], {
   productionBrowserSourceMaps: true,
   serverRuntimeConfig: {
     rootDir: __dirname,
   },
   webpack: (config, options) => {
-    // In `pages/_app.js`, Sentry is imported from @sentry/browser. While
-    // @sentry/node will run in a Node.js environment. @sentry/node will use
-    // Node.js-only APIs to catch even more unhandled exceptions.
-    //
-    // This works well when Next.js is SSRing your page on a server with
-    // Node.js, but it is not what we want when your client-side bundle is being
-    // executed by a browser.
-    //
-    // Luckily, Next.js will call this webpack function twice, once for the
-    // server and once for the client. Read more:
-    // https://nextjs.org/docs/api-reference/next.config.js/custom-webpack-config
-    //
-    // So ask Webpack to replace @sentry/node imports with @sentry/browser when
-    // building the browser's bundle
-    if (!options.isServer) {
-      config.resolve.alias["@sentry/node"] = "@sentry/browser";
-    }
-
-    // config.node = {
-    //   fs: 'empty',
-    // };
-
-    // When all the Sentry configuration env variables are available/configured
-    // The Sentry webpack plugin gets pushed to the webpack plugins to build
-    // and upload the source maps to sentry.
-    // This is an alternative to manually uploading the source maps
-    // Note: This is disabled in development mode.
-    if (
-      SENTRY_DSN &&
-      SENTRY_ORG &&
-      SENTRY_PROJECT &&
-      SENTRY_AUTH_TOKEN &&
-      COMMIT_SHA &&
-      NODE_ENV === "production"
-    ) {
-      config.plugins.push(
-        new SentryWebpackPlugin({
-          include: ".next",
-          ignore: ["node_modules"],
-          stripPrefix: ["webpack://_N_E/"],
-          urlPrefix: `~${basePath}/_next`,
-          release: COMMIT_SHA,
-        })
-      );
-    }
     return config;
   },
   basePath,
@@ -133,3 +66,19 @@ module.exports = withPlugins([[withBundleAnalyzer]], {
     domains: ["cdn.plant-for-the-planet.org", "cdn.planetapp.workers.dev"],
   },
 });
+
+const SentryWebpackPluginOptions = {
+  // Additional config options for the Sentry Webpack plugin. Keep in mind that
+  // the following options are set automatically, and overriding them is not
+  // recommended:
+  //   release, url, org, project, authToken, configFile, stripPrefix,
+  //   urlPrefix, include, ignore
+
+  silent: true, // Suppresses all logs
+  // For all available options, see:
+  // https://github.com/getsentry/sentry-webpack-plugin#options.
+};
+
+// Make sure adding Sentry options is the last code to run before exporting, to
+// ensure that your source maps include changes from all other Webpack plugins
+module.exports = withSentryConfig(moduleExports, SentryWebpackPluginOptions);
