+++ package.json
@@ -19,6 +19,7 @@
     "@paypal/react-paypal-js": "^7.5.0",
     "@sentry/browser": "^6.2.5",
     "@sentry/integrations": "^6.2.5",
+    "@sentry/nextjs": "^6.16.1",
     "@sentry/node": "^6.2.5",
     "@sentry/webpack-plugin": "^1.14.2",
     "@stripe/react-stripe-js": "^1.4.0",
