+++ pages/_app.tsx
@@ -1,9 +1,6 @@
 import type { AppProps /*, AppContext */ } from "next/app";
 import Layout from "../src/Layout";
 import ThemeProvider, { useTheme } from "../styles/themeContext";
-import getConfig from "next/config";
-import * as Sentry from "@sentry/node";
-import { RewriteFrames } from "@sentry/integrations";
 import theme from "./../styles/theme";
 import "./../styles/globals.scss";
 import "./../styles/footer.scss";
@@ -18,48 +15,6 @@ import { useRouter } from "next/router";
 import React from "react";
 import { browserNotCompatible } from "../src/Utils/browsercheck";
 import BrowserNotSupported from "./../src/Common/ContentLoaders/BrowserNotSupported";
-if (process.env.NEXT_PUBLIC_SENTRY_DSN) {
-  const config = getConfig();
-  const distDir = `${config.serverRuntimeConfig.rootDir}/.next`;
-  Sentry.init({
-    enabled: process.env.NODE_ENV === "production",
-    integrations: [
-      new RewriteFrames({
-        iteratee: (frame) => {
-          frame.filename = frame.filename?.replace(distDir, "app:///_next");
-          return frame;
-        },
-      }),
-    ],
-    dsn: process.env.NEXT_PUBLIC_SENTRY_DSN,
-    // from https://gist.github.com/pioug/b006c983538538066ea871d299d8e8bc,
-    // also see https://docs.sentry.io/platforms/javascript/configuration/filtering/#decluttering-sentry
-    ignoreErrors: [
-      /^No error$/,
-      /__show__deepen/,
-      /_avast_submit/,
-      /Access is denied/,
-      /anonymous function: captureException/,
-      /Blocked a frame with origin/,
-      /console is not defined/,
-      /cordova/,
-      /DataCloneError/,
-      /Error: AccessDeny/,
-      /event is not defined/,
-      /feedConf/,
-      /ibFindAllVideos/,
-      /myGloFrameList/,
-      /SecurityError/,
-      /MyIPhoneApp/,
-      /snapchat.com/,
-      /vid_mate_check is not defined/,
-      /win\.document\.body/,
-      /window\._sharedData\.entry_data/,
-      /ztePageScrollModule/,
-    ],
-    denyUrls: [],
-  });
-}
 
 function MyApp({ Component, pageProps }: AppProps) {
   const router = useRouter();
