+++ sentry.properties
@@ -1,0 +1,4 @@
+defaults.url=https://sentry.io/
+defaults.org=plant-for-the-planet
+defaults.project=planet-donations
+# cli.executable=../path/to/bin/sentry-cli
